added payment table in mySQL schema and added relevant informations
